pyproject.toml
[project]
name = "sjphy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cycler>=0.12.1",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "rich-gradient>=0.3.6",
    "uncertainties>=3.2.3",
    "scipy>=1.15.3",
    "poethepoet>=0.37.0",
    "pyperclip>=1.11.0",
]

[tool.setuptools]
package-dir = {"" = "lib"}
packages = ["graphing", "me"]

[tool.poe.tasks]
sync-deps = "uv sync"
sync-lib = "uv pip install -e ."
main = "uv run main.py"
sync = { sequence = ["sync-deps", "sync-lib", "main"], help = "Syncs dependencies and the local library in editable mode." }
